Add several behaviors of ANT controlpad
