+++ app-launcher/controlpad/index.html
@@ -23,6 +23,7 @@
 
   <script type="text/javascript" src="./createapp-dialog-view.js"></script>
   <script type="text/javascript" src="./snackbar-view.js"></script>
+  <script type="text/javascript" src="./createapp-fabbutton-view.js"></script>
 
   <script type="text/javascript" src="./navmenu-view.js"></script>
   <script type="text/javascript" src="./appselector-label-view.js"></script>
@@ -92,6 +93,11 @@
       </div>
     </main>
   </div>
+
+  <!-- Create app fab button -->
+  <button id="createapp-fabbutton" class="mdl-button mdl-js-button mdl-button--colored mdl-button--fab mdl-js-ripple-effect" style="position: fixed; display: block; right: 0; bottom: 0; margin-right: 40px; margin-bottom: 40px; z-index: 900;">
+    <i class="material-icons">add</i>
+  </button>
 
   <!-- Create app dialog -->
   <dialog id="create-app-dialog" class="mdl-dialog">
